TST: ci issue, fixed errors in cdi part

+++ skbeam/core/cdi.py
@@ -62,7 +62,7 @@ def _dist(dims):
         of the array of shape `dims`
     """
     dist_sum = []
-    shape = np.ones(len(dims))
+    shape = np.ones(len(dims), dtype=np.int)
     for idx, d in enumerate(dims):
         vec = (np.arange(d) - d // 2) ** 2
         shape[idx] = -1
